more phpstan fixes
