+++ src/Viserio/Component/Routing/Tests/Traits/MiddlewareAwareTraitTest.php
@@ -3,17 +3,16 @@ declare(strict_types=1);
 namespace Viserio\Component\Routing\Tests\Traits;
 
 use PHPUnit\Framework\TestCase;
-use Viserio\Component\Contract\Routing\MiddlewareAware as MiddlewareAwareContract;
 use Viserio\Component\Routing\Tests\Fixture\FakeMiddleware;
 use Viserio\Component\Routing\Tests\Fixture\FooMiddleware;
-use Viserio\Component\Routing\Traits\MiddlewareAwareTrait;
+use Viserio\Component\Routing\Tests\Fixture\MiddlewareHandler;
 
 class MiddlewareAwareTraitTest extends TestCase
 {
     public function testWithMiddlewareObject(): void
     {
         $middleware = new FooMiddleware();
-        $object     = $this->getMiddlewareAwareObject(true);
+        $object     = new MiddlewareHandler(true);
 
         $object->withMiddleware($middleware);
 
@@ -22,7 +21,7 @@ class MiddlewareAwareTraitTest extends TestCase
 
     public function testWithMiddlewareString(): void
     {
-        $object = $this->getMiddlewareAwareObject(true);
+        $object = new MiddlewareHandler(true);
 
         $object->withMiddleware(FooMiddleware::class);
 
@@ -31,7 +30,7 @@ class MiddlewareAwareTraitTest extends TestCase
 
     public function testWithMiddlewareArray(): void
     {
-        $object = $this->getMiddlewareAwareObject(true);
+        $object = new MiddlewareHandler(true);
 
         $object->withMiddleware([FooMiddleware::class, FakeMiddleware::class]);
 
@@ -40,7 +39,7 @@ class MiddlewareAwareTraitTest extends TestCase
 
     public function testWithoutMiddlewareWithString(): void
     {
-        $object = $this->getMiddlewareAwareObject(true, true);
+        $object = new MiddlewareHandler(true, true);
 
         $object->withoutMiddleware(FooMiddleware::class);
 
@@ -49,7 +48,7 @@ class MiddlewareAwareTraitTest extends TestCase
 
     public function testWithoutMiddlewareWithArray(): void
     {
-        $object = $this->getMiddlewareAwareObject(true, true);
+        $object = new MiddlewareHandler(true, true);
 
         $object->withoutMiddleware([FooMiddleware::class, FooMiddleware::class]);
 
@@ -58,7 +57,7 @@ class MiddlewareAwareTraitTest extends TestCase
 
     public function testWithoutMiddlewareWithNull(): void
     {
-        $object = $this->getMiddlewareAwareObject(true, true);
+        $object = new MiddlewareHandler(true, true);
 
         $object->withMiddleware(FooMiddleware::class);
         $object->withoutMiddleware(null);
@@ -69,13 +68,13 @@ class MiddlewareAwareTraitTest extends TestCase
 
     public function testAliasMiddleware(): void
     {
-        $object = $this->getMiddlewareAwareObject(true);
+        $object = new MiddlewareHandler(true);
         $object->aliasMiddleware('foo', FooMiddleware::class);
 
         self::assertSame(['foo' => FooMiddleware::class], $object->getMiddlewares());
 
         $middleware = new FooMiddleware();
-        $object     = $this->getMiddlewareAwareObject(true);
+        $object     = new MiddlewareHandler(true);
 
         $object->aliasMiddleware('bar', $middleware);
 
@@ -88,7 +87,7 @@ class MiddlewareAwareTraitTest extends TestCase
      */
     public function testAliasMiddlewareThrowException(): void
     {
-        $object = $this->getMiddlewareAwareObject(true);
+        $object = new MiddlewareHandler(true);
 
         $object->aliasMiddleware('foo', FooMiddleware::class);
         $object->aliasMiddleware('foo', FooMiddleware::class);
@@ -100,7 +99,7 @@ class MiddlewareAwareTraitTest extends TestCase
      */
     public function testAliasMiddlewareThrowExceptionWithWrongType(): void
     {
-        $this->getMiddlewareAwareObject(true)->aliasMiddleware('foo', null);
+        (new MiddlewareHandler(true))->aliasMiddleware('foo', null);
     }
 
     /**
@@ -109,7 +108,7 @@ class MiddlewareAwareTraitTest extends TestCase
      */
     public function testWithWrongMiddleware(): void
     {
-        $this->getMiddlewareAwareObject(true, true)->withMiddleware(MiddlewareAwareTraitTest::class);
+        (new MiddlewareHandler(true, true))->withMiddleware(MiddlewareAwareTraitTest::class);
     }
 
     /**
@@ -118,46 +117,6 @@ class MiddlewareAwareTraitTest extends TestCase
      */
     public function testWithWrongType(): void
     {
-        $this->getMiddlewareAwareObject(true)->withMiddleware(null);
-    }
-
-    /**
-     * @param bool $resetMiddlewares
-     * @param bool $resetBypassedMiddlewares
-     *
-     * @return \Viserio\Component\Contract\Routing\MiddlewareAware
-     */
-    private function getMiddlewareAwareObject(bool $resetMiddlewares = false, bool $resetBypassedMiddlewares = false): MiddlewareAwareContract
-    {
-        return new class($resetMiddlewares, $resetBypassedMiddlewares) implements MiddlewareAwareContract {
-            use MiddlewareAwareTrait;
-
-            public function __construct($resetMiddlewares, $resetBypassedMiddlewares)
-            {
-                if ($resetMiddlewares) {
-                    $this->middlewares = [];
-                }
-
-                if ($resetBypassedMiddlewares) {
-                    $this->bypassedMiddlewares = [];
-                }
-            }
-
-            /**
-             * @return array
-             */
-            public function getMiddlewares(): array
-            {
-                return $this->middlewares;
-            }
-
-            /**
-             * @return array
-             */
-            public function getBypassedMiddlewares(): array
-            {
-                return $this->bypassedMiddlewares;
-            }
-        };
+        (new MiddlewareHandler(true))->withMiddleware(null);
     }
 }
